Revert "rename account_entry"
This reverts commit 78ff33bf376d9752c5f2e15bef137e8c93f1029d.
