+++ state/src/state_object_entry.rs
